Ignores array types as far as Migrator is concerned
Previously, a ecto type like {:array, :string} or {:array, :map} would
be fed into Code.ensure_loaded?/1, raising an argument error. While
it would be nice to be able to handle array types with encoded members,
ecto does not seem to handle the encoding to postgres nicely, and having
postgres report the field as an array of binaries leaks more information
than just having a binary type for the field and using Cloak.Ecto.Map to
wrap the data in a way that can be turned back into the list you want by
the app.