Remove pbkdf2 optional dependency
It isn't possible to publish a package to Hex.pm with a dependency on a
Github fork which is not also hosted on Hex.pm. The current version of
`pbkdf2` is not compatible with Erlang 24. To use it, you'll need a
forked version, see #24.

For this reason, I needed to make the `pbkdf2` a `:dev, :test`
dependency instead of an `:optional` dependency for `cloak_ecto`. If you
add it to your mix dependencies in your project, `cloak_ecto` should
still detect it; it shouldn't need to be listed as an optional
dependency.