:heavy_minus_sign: Remove dependency on PBKDF2
This is a breaking change. OTP versions > 24.2 now include the `pbkdf2_hmac/5`
function in the `:crypto` module, thanks to this PR:
https://github.com/erlang/otp/pull/5421

The new function does not include support for all the algorithms that the old
dependency supported, namely `md4`, `md5` and `ripemd`. However, I think this
is a valid tradeoff.

It is no longer necessary to add a `:pbkdf2` dependency in order to use the
PBKDF2 field type.